+++ lib/fossa/settings/global_config.py
@@ -1,7 +1,7 @@
 class BaseConfig:
     DEBUG = False
     APP_TITLE = "Fossa"
-    PREFERRED_URL_SCHEME = "https"
+    PREFERRED_URL_SCHEME = "http"
     SECRET_KEY = ""
     HTTP_PORT = 2345
     ACCEPTED_MODEL_CLASSES = []  # iterable of classes that the node is authorised to run